@@ -6,25 +6,33 @@
 	"scripts": {
 		"dev": "vite dev",
 		"prod": "node ./server.cjs",
-		"build": "vite build"
+		"build": "vite build",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"lint": "eslint ./src"
 	},
 	"engines": {
 		"node": ">=21.0"
 	},
 	"devDependencies": {
-		"@sveltejs/vite-plugin-svelte": "^3.0",
-		"prettier": "^3.2.4",
+		"@eslint/js": "^9.33.0",
+		"@sveltejs/vite-plugin-svelte": "^5.0.0",
+		"@tsconfig/svelte": "^5.0.2",
+		"@types/node": "^22.0",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.11.0",
+		"globals": "^16.3",
+		"prettier": "^3.3",
 		"prettier-plugin-svelte": "^3.2.1",
 		"sass": "^1.62.1",
-		"svelte": "^4.0",
-		"svelte-check": "^3.6.2",
-		"svelte-preprocess": "^5.0.3",
-		"typescript-svelte-plugin": "^0.3.45",
-		"vite": "^5.0"
+		"svelte-check": "^4.0",
+		"svelte-preprocess": "^6.0",
+		"tslib": "^2.6.2",
+		"typescript": "^5.2.2",
+		"typescript-eslint": "^8.39.1",
+		"vite": "^6.0.0"
 	},
 	"dependencies": {
-		"crelte": "^0.4.0",
-		"crelte-node": "^0.4.0",
-		"crelte-std": "^0.1.1"
+		"svelte": "^5.0",
+		"crelte": "^0.5.5"
 	}
 }
